+++ src/theme.ts
@@ -6,7 +6,8 @@ import merge from 'deepmerge'
 import { DarkTheme as PaperDarkTheme } from 'react-native-paper'
 import { Theme as PaperTheme } from 'react-native-paper/lib/typescript/types'
 
-export type Theme = PaperTheme & NavigationTheme & { colors: { card: string } }
+export type Theme = PaperTheme &
+	NavigationTheme & { colors: { card: string; success: string } }
 
 const mergedDefaultThemes = merge(NavigationDarkTheme, PaperDarkTheme)
 
@@ -17,6 +18,7 @@ export const theme: Theme = {
 		...mergedDefaultThemes.colors,
 		background: '#292d3d',
 		primary: '#AF254B',
+		success: 'green',
 		text: '#FFFFFF',
 		notification: '#252939',
 		card: '#0C1A31',
