+++ translations/es.json
@@ -52,5 +52,14 @@
 		"assignment": "Asignar variable",
 		"selectVariable": "Seleccionar variable",
 		"messageSend": "Enviar mensaje"
+	},
+	"tests": {
+		"title": "Grupos de tests",
+		"describeName": "Nombre del grupo",
+		"newTest": "Crear test",
+		"testName": "Nombre del test"
+	},
+	"project": {
+		"title": "Mi Proyecto"
 	}
 }
